backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
harperdb==0.4.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
